Refactor: move dossiers and organes fetching functions to repondeur
