+++ repondeur/zam_repondeur/data.py
@@ -1,11 +1,14 @@
 from pyramid.config import Configurator
 from pyramid.threadlocal import get_current_registry
 
-from zam_aspirateur.textes.dossiers_legislatifs import get_dossiers_legislatifs
-from zam_aspirateur.textes.organes import get_organes
+from zam_repondeur.fetch.an.dossiers.dossiers_legislatifs import (
+    get_dossiers_legislatifs
+)
+from zam_repondeur.fetch.an.organes import get_organes
 
 
 def load_data(config: Configurator) -> None:
+
     data = config.registry.settings.setdefault("data", {})
     current_legislature = int(config.registry.settings["zam.legislature"])
     data["dossiers"] = get_dossiers_legislatifs(legislature=current_legislature)
